[ticket/12331] Fix SQL error when inserting a new entry to profile field data
There was a "Field 'pf_phpbb_interests' doesn't have a default value" error,
because we didn't specify values for the disabled fields.
We can not disable the sql_return_on_error() thou, because it otherwise we run
into "Duplicate entry '2' for key 'PRIMARY'" errors when the user submits the
form again but didn't change any values.

PHPBB3-12331