[ticket/15320] 	Redis cache does not save keys withouth expiration
In some functions like sql_save in cache/memory.php, code try to save a key/value in cache with ttl = 0 so key should never expire. In current redis.php cache driver, it fails so key never get cached. This cause for example that when you create a subforum, it not appear in the forum box in the admincp.
To solve, if ttl is 0, we use redis->set instead of setex