[ticket/16556] Remove purposeless <Files "*"> from .htaccess
Wrapping access permissions into <Files "*"> at best has zero effect
and implies unnecessary code and parsing for the web server. At least
it does not block access to files only, but still denies auto indexing
and access to sub directories effectively as well. But removing this
directive is still the cleaner and safer way to deny access to any
kind of resource that is provided within the directory in question.

To deny access to migration data, a single .htaccess file can be used.
This reduces the effort for future changes and it is not required
anymore to create new .htaccess files for every new migration
directory.

Additionally this corrects the fact the "Require" is part of
"mod_authz_core", not "mod_authz_host".

PHPBB3-16556

Signed-off-by: MichaIng <micha@dietpi.com>